requirements.txt
# Requirements for Whisper to CoreML conversion
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
coremltools>=7.0
openai-whisper>=20231117
ane_transformers>=0.1.0
transformers>=4.18.0
numpy>=1.20.0
